requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.2.0
authy==2.2.6
billiard==4.2.0
bitarray==2.9.2
boto3==1.35.24
botocore==1.35.24
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
ckzg==2.0.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==43.0.1
cytoolz==0.12.3
dj-rest-auth==6.0.0
Django==5.1.1
django-allauth==64.2.1
django-cors-headers==4.4.0
django-environ==0.11.2
django-formtools==2.5.1
django-otp==1.5.4
django-phonenumber-field==8.0.0
django-storages==1.14.4
django-two-factor-auth==1.17.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
easy-colors==0.0.1
eth-account==0.13.3
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==2.1.0
eth-typing==5.0.0
eth-utils==5.0.0
eth_abi==5.1.0
frozenlist==1.4.1
gunicorn==23.0.0
hexbytes==1.2.1
idna==3.10
inflection==0.5.1
jmespath==1.0.1
kombu==5.4.2
multidict==6.1.0
packaging==24.1
parsimonious==0.10.0
pillow==10.4.0
prompt_toolkit==3.0.47
psycopg2==2.9.9
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pypng==0.20220715.0
python-dateutil==2.9.0.post0
pytz==2024.2
pyunormalize==16.0.0
pywin32==306
PyYAML==6.0.2
qrcode==7.4.2
regex==2024.9.11
requests==2.32.3
rlp==4.0.1
s3transfer==0.10.2
send-mail==0.0.2
six==1.16.0
sqlparse==0.5.1
toolz==0.12.1
twilio==9.3.1
types-requests==2.32.0.20240914
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
web3==7.2.0
websockets==13.0.1
yarl==1.11.1
